+++ src/main/java/com/example/automarket/domain/DealType.java
@@ -1,0 +1,6 @@
+package com.example.automarket.domain;
+
+public enum DealType {
+    BUY,
+    SELL
+}
