--- src/main/java/com/example/automarket/domain/model/DealType.java
@@ -1,6 +1,0 @@
-package com.example.automarket.domain.model;
-
-public enum DealType {
-    BUY,
-    SELL
-}
