+++ src/main/java/com/example/automarket/domain/model/listing/vehicle/TrailerListing.java
@@ -1,5 +1,6 @@
-package com.example.automarket.domain.model.listing;
+package com.example.automarket.domain.model.listing.vehicle;
 
+import com.example.automarket.domain.model.listing.Loadable;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
