+++ src/main/java/com/example/automarket/repository/PartRepository.java
@@ -1,6 +1,6 @@
 package com.example.automarket.repository;
 
-import com.example.automarket.domain.model.listing.PartListing;
+import com.example.automarket.domain.model.listing.part.PartListing;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
