+++ src/main/java/com/example/automarket/repository/VehicleRepository.java
@@ -1,6 +1,6 @@
 package com.example.automarket.repository;
 
-import com.example.automarket.domain.model.listing.VehicleListing;
+import com.example.automarket.domain.model.listing.vehicle.VehicleListing;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
