+++ src/main/java/com/example/automarket/service/impl/VehicleServiceImpl.java
@@ -3,10 +3,10 @@ package com.example.automarket.service.impl;
 import com.example.automarket.domain.dto.response.CarListingResponse;
 import com.example.automarket.domain.dto.response.TrailerListingResponse;
 import com.example.automarket.domain.dto.response.VehicleListingResponse;
-import com.example.automarket.domain.model.listing.CarListing;
 import com.example.automarket.domain.model.listing.Listing;
-import com.example.automarket.domain.model.listing.TrailerListing;
-import com.example.automarket.domain.model.listing.VehicleListing;
+import com.example.automarket.domain.model.listing.vehicle.CarListing;
+import com.example.automarket.domain.model.listing.vehicle.TrailerListing;
+import com.example.automarket.domain.model.listing.vehicle.VehicleListing;
 import com.example.automarket.repository.VehicleRepository;
 import com.example.automarket.service.VehicleService;
 import lombok.RequiredArgsConstructor;
